+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/app.rs
@@ -7,6 +7,27 @@ use types::CaseMatching;
 
 use crate::command;
 
+/// This command is only invoked when user uses the prebuilt binary, more specifically, exe in
+/// vim-clap/bin/maple.
+#[derive(Parser, Debug, Clone)]
+pub struct Upgrade {
+    /// Download if the local version mismatches the latest remote version.
+    #[clap(long)]
+    download: bool,
+    /// Disable the downloading progress_bar
+    #[clap(long)]
+    no_progress_bar: bool,
+}
+
+impl Upgrade {
+    pub async fn run(self, local_tag: &str) -> Result<()> {
+        upgrade::Upgrade::new(self.download, self.no_progress_bar)
+            .run(local_tag)
+            .await
+            .map_err(Into::into)
+    }
+}
+
 #[derive(Parser, Debug)]
 pub enum Cmd {
     /// Display the current version
@@ -46,12 +67,12 @@ pub enum Cmd {
     RipGrepForerunner(command::grep::RipGrepForerunner),
     /// Retrive the latest remote release info.
     #[clap(name = "upgrade")]
-    Upgrade(upgrade::Upgrade),
+    Upgrade(Upgrade),
 }
 
 #[derive(Parser, Debug)]
 #[clap(name = "maple")]
-#[clap(global_setting(AppSettings::DisableVersionFlag))]
+#[clap(global_setting(AppSettings::NoAutoVersion))]
 pub struct Maple {
     #[clap(flatten)]
     pub params: Params,
@@ -116,14 +137,41 @@ impl Maple {
             Cmd::DumbJump(dumb_jump) => dumb_jump.run().await?,
             Cmd::RipGrepForerunner(rip_grep_forerunner) => rip_grep_forerunner.run(self.params)?,
             Cmd::Rpc => {
-                if let Some(ref log_path) = self.log {
-                    crate::logger::init(log_path)?;
-                } else if let Ok(log_path) = std::env::var("VIM_CLAP_LOG_PATH") {
-                    crate::logger::init(log_path)?;
-                }
+                let maybe_log = if let Some(log_path) = self.log {
+                    Some(log_path)
+                } else if let Ok(log_path) =
+                    std::env::var("VIM_CLAP_LOG_PATH").map(std::path::PathBuf::from)
+                {
+                    Some(log_path)
+                } else {
+                    None
+                };
 
-                crate::stdio_server::run_forever(std::io::BufReader::new(std::io::stdin()));
-                // crate::stdio_server::start()?;
+                if let Some(log_path) = maybe_log {
+                    if let Ok(metadata) = std::fs::metadata(&log_path) {
+                        if log_path.is_file() && metadata.len() > 8 * 1024 * 1024 {
+                            std::fs::remove_file(&log_path)?;
+                        }
+                    }
+
+                    let file_name = log_path.file_name().expect("Invalid file name");
+                    let directory = log_path.parent().expect("A file must have a parent");
+
+                    let file_appender = tracing_appender::rolling::never(directory, file_name);
+                    let (non_blocking, _guard) = tracing_appender::non_blocking(file_appender);
+
+                    let subscriber = tracing_subscriber::FmtSubscriber::builder()
+                        .with_max_level(tracing::Level::TRACE)
+                        .with_line_number(true)
+                        .with_writer(non_blocking)
+                        .finish();
+
+                    tracing::subscriber::set_global_default(subscriber)?;
+
+                    crate::stdio_server::run_forever(std::io::BufReader::new(std::io::stdin()));
+                } else {
+                    crate::stdio_server::run_forever(std::io::BufReader::new(std::io::stdin()));
+                }
             }
         };
         Ok(())