+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/command/filter.rs
@@ -4,14 +4,12 @@ use std::path::PathBuf;
 use anyhow::Result;
 use clap::Parser;
 
-use filter::{
-    matcher::{Bonus, FuzzyAlgorithm, Matcher, MatchingTextKind},
-    subprocess::Exec,
-    FilterContext, Source,
-};
+use filter::{subprocess::Exec, FilterContext, Source};
+use matcher::{Bonus, FuzzyAlgorithm, MatchScope, Matcher};
 use types::SourceItem;
 
-use crate::{app::Params, paths::AbsPathBuf};
+use crate::app::Params;
+use crate::paths::AbsPathBuf;
 
 fn parse_bonus(s: &str) -> Bonus {
     if s.to_lowercase().as_str() == "filename" {
@@ -50,7 +48,7 @@ pub struct Filter {
 
     /// Apply the filter on the full line content or parial of it.
     #[clap(long, parse(from_str), default_value = "full")]
-    matching_text_kind: MatchingTextKind,
+    match_scope: MatchScope,
 
     /// Add a bonus to the score of base matching algorithm.
     #[clap(long, parse(from_str = parse_bonus), default_value = "none")]
@@ -106,7 +104,7 @@ impl Filter {
             ..
         }: Params,
     ) -> Result<()> {
-        let matcher = Matcher::with_bonuses(self.get_bonuses(), self.algo, self.matching_text_kind)
+        let matcher = Matcher::with_bonuses(self.get_bonuses(), self.algo, self.match_scope)
             .set_case_matching(case_matching);
 
         if self.sync {