+++ .vim/pack/basic/start/vim-clap/crates/maple_cli/src/utils.rs
@@ -1,11 +1,11 @@
-use std::{
-    io::{BufRead, Lines},
-    path::{Path, PathBuf},
-};
+use std::borrow::Cow;
+use std::io::{BufRead, Lines};
+use std::path::{Path, PathBuf};
 
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use chrono::prelude::*;
 use directories::ProjectDirs;
+use itertools::Itertools;
 use once_cell::sync::Lazy;
 
 use filter::subprocess::Exec;
@@ -19,6 +19,13 @@ use utility::{println_json, println_json_with_length, read_first_lines};
 pub static PROJECT_DIRS: Lazy<ProjectDirs> = Lazy::new(|| {
     ProjectDirs::from("org", "vim", "Vim Clap")
         .expect("Couldn't create project directory for vim-clap")
+});
+
+pub static HOME_DIR: Lazy<PathBuf> = Lazy::new(|| {
+    directories::BaseDirs::new()
+        .expect("Failed to construct BaseDirs")
+        .home_dir()
+        .to_path_buf()
 });
 
 /// Yes or no terms.
@@ -171,10 +178,7 @@ pub(crate) fn expand_tilde(path: impl AsRef<str>) -> Result<PathBuf> {
     static HOME_PREFIX: Lazy<String> = Lazy::new(|| format!("~{}", std::path::MAIN_SEPARATOR));
 
     let fpath = if let Some(stripped) = path.as_ref().strip_prefix(HOME_PREFIX.as_str()) {
-        let mut home_dir = directories::BaseDirs::new()
-            .ok_or_else(|| anyhow!("Failed to construct BaseDirs"))?
-            .home_dir()
-            .to_path_buf();
+        let mut home_dir = HOME_DIR.clone();
         home_dir.push(stripped);
         home_dir
     } else {
@@ -241,6 +245,66 @@ pub fn display_width(n: usize) -> usize {
     len
 }
 
+// /home/xlc/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
+pub(crate) fn truncate_absolute_path(abs_path: &str, max_len: usize) -> Cow<'_, str> {
+    if abs_path.len() > max_len {
+        let gap = abs_path.len() - max_len;
+
+        const SEP: char = std::path::MAIN_SEPARATOR;
+
+        if let Some(home_dir) = crate::utils::HOME_DIR.as_path().to_str() {
+            if abs_path.starts_with(home_dir) {
+                // ~/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
+                if home_dir.len() > gap {
+                    return abs_path.replacen(home_dir, "~", 1).into();
+                }
+
+                // ~/.rustup/.../github.com/paritytech/substrate/frame/system/src/lib.rs
+                let home_stripped = &abs_path.trim_start_matches(home_dir)[1..];
+                if let Some((first, target)) = home_stripped.split_once(SEP) {
+                    let mut hidden = 0usize;
+                    for component in target.split(SEP) {
+                        if hidden > gap + 2 {
+                            let mut target = target.to_string();
+                            target.replace_range(..hidden - 1, "...");
+                            return format!("~{SEP}{first}{SEP}{target}").into();
+                        } else {
+                            hidden += component.len() + 1;
+                        }
+                    }
+                }
+            } else {
+                let top = abs_path.splitn(4, SEP).collect::<Vec<_>>();
+                if let Some(last) = top.last() {
+                    if let Some((_first, target)) = last.split_once(SEP) {
+                        let mut hidden = 0usize;
+                        for component in target.split(SEP) {
+                            if hidden > gap + 2 {
+                                let mut target = target.to_string();
+                                target.replace_range(..hidden - 1, "...");
+                                let head = top.iter().take(top.len() - 1).join(&SEP.to_string());
+                                return format!("{head}{SEP}{target}").into();
+                            } else {
+                                hidden += component.len() + 1;
+                            }
+                        }
+                    }
+                }
+            }
+        } else {
+            // Truncate the left of absolute path string.
+            // ../stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
+            if let Some((offset, _)) = abs_path.char_indices().nth(abs_path.len() - max_len + 2) {
+                let mut abs_path = abs_path.to_string();
+                abs_path.replace_range(..offset, "..");
+                return abs_path.into();
+            }
+        }
+    }
+
+    abs_path.into()
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -249,5 +313,29 @@ mod tests {
     fn test_count_lines() {
         let f: &[u8] = b"some text\nwith\nfour\nlines\n";
         assert_eq!(count_lines(f).unwrap(), 4);
+    }
+
+    #[test]
+    fn test_truncate_absolute_path() {
+        #[cfg(not(target_os = "windows"))]
+        let p = ".rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs";
+        #[cfg(target_os = "windows")]
+        let p = r#".rustup\toolchains\stable-x86_64-unknown-linux-gnu\lib\rustlib\src\rust\library\alloc\src\string.rs"#;
+        let abs_path = format!(
+            "{}{}{}",
+            crate::utils::HOME_DIR.as_path().to_str().unwrap(),
+            std::path::MAIN_SEPARATOR,
+            p
+        );
+        let max_len = 60;
+        #[cfg(not(target_os = "windows"))]
+        let expected = "~/.rustup/.../src/rust/library/alloc/src/string.rs";
+        #[cfg(target_os = "windows")]
+        let expected = r#"~\.rustup\...\src\rust\library\alloc\src\string.rs"#;
+        assert_eq!(truncate_absolute_path(&abs_path, max_len), expected);
+
+        let abs_path = "/media/xlc/Data/src/github.com/paritytech/substrate/bin/node/cli/src/command_helper.rs";
+        let expected = "/media/xlc/.../bin/node/cli/src/command_helper.rs";
+        assert_eq!(truncate_absolute_path(abs_path, max_len), expected);
     }
 }